Rename shader mods to shaders
