+++ src/render/pipeline/glow/mod.rs
@@ -1,4 +1,4 @@
-pub mod shader;
+pub mod shaders;
 
 use log::info;
 
@@ -46,7 +46,7 @@ impl ScenePassComponent for Glow {
         &self,
         core: render::shader::Core<(Context, P), V>,
     ) -> render::shader::Core<(Context, P), V> {
-        shader::glow_map_core_transform(core)
+        shaders::glow_map_core_transform(core)
     }
 
     fn output_textures(&self) -> Vec<(&'static str, &glium::texture::Texture2d)> {
@@ -59,7 +59,7 @@ impl CompositionPassComponent for Glow {
         &self,
         core: render::shader::Core<(), screen_quad::Vertex>,
     ) -> render::shader::Core<(), screen_quad::Vertex> {
-        shader::composition_core_transform(core)
+        shaders::composition_core_transform(core)
     }
 }
 
@@ -74,7 +74,7 @@ impl Glow {
         let glow_texture_back = Self::create_texture(facade, rounded_size)?;
 
         info!("Creating blur program");
-        let blur_program = shader::blur_core().build_program(facade)?;
+        let blur_program = shaders::blur_core().build_program(facade)?;
 
         info!("Creating screen quad");
         let screen_quad = ScreenQuad::create(facade)?;
